+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,13 @@
 # sit-onyx
 
+## 1.0.0-beta.220
+
+### Minor Changes
+
+- fc57919: - feat(OnyxDataGrid): implemented support for passing column type specific configuration
+  - all default column types now support `options` for overwriting the formatting
+  - feat(i18n): allow passing INTL formatting options for number `n` and date `d`formatter
+
 ## 1.0.0-beta.219
 
 ### Minor Changes